Fixed makefile wildcard syntax to work on all shells
The previous makefile used a glob pattern to select all C files to be compiled
(`LIBSOURCES=*.c`). This properly expands when the IDE launches the make process
in bash, but fails to expand on shells such as CMD.EXE. Using `$(wildcard *.c)`
is more reliable, as the expansion is done by make, not the shell, and thus will
work on any platform.

It also seems like most other IPs use this method, except for MIPI_CSI_2_RX_v1_0
and MIPI_D_PHY_RX_v1_0, but in those cases the expanded pattern is never used so
this error doesn't result in the driver not building.